Flytt ExceptionsHandler i JournalpostExceptionHandler til ApiExceptionHandler
Da slipper man å måtte tenkte på @Order og konkurrerende handlers. Er dessuten ryddig å ha alt på ett sted
